change np float to python builtin

+++ spidet/domain/Trace.py
@@ -20,11 +20,11 @@ class Trace:
     start_timestamp: float
         The start timestamp of the recording as a UNIX timestamp.
 
-    data: numpy.ndarray[numpy.dtype[numpy.float64]]
+    data: numpy.ndarray[numpy.dtype[float]]
         An array containing the EEG data of the given channel.
     """
 
     label: str
     sfreq: int
     start_timestamp: float
-    data: np.ndarray[np.dtype[np.float64]]
+    data: np.ndarray[np.dtype[float]]
