+++ spidet/spike_detection/spike_detection_pipeline.py
@@ -58,7 +58,7 @@ class SpikeDetectionPipeline:
         A floating point number :math:`\in [0, 1]`.
         If this parameter is non-zero, nonnegative matrix factorization is run with sparseness constraints.
 
-    bad_times: numpy.ndarray[numpy.dtype[numpy.float64]]
+    bad_times: numpy.ndarray[numpy.dtype[float]]
         An optional N x 2 numpy array containing periods that must be excluded before applying
         the line-length transformation. Each of th N rows in the array represents a period to be excluded,
         defined by the start and end indices of the period in the original iEEG data.
@@ -82,7 +82,7 @@ class SpikeDetectionPipeline:
         results_dir: str = None,
         save_nmf_matrices: bool = False,
         sparseness: float = 0.0,
-        bad_times: np.ndarray[np.dtype[np.float64]] = None,
+        bad_times: np.ndarray[np.dtype[float]] = None,
         nmf_runs: int = 100,
         rank_range: Tuple[int, int] = (2, 5),
         line_length_freq: int = 50,
